+++ src/main/scala-2.11/markets/tradables/Quantity.scala
@@ -13,9 +13,7 @@ WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
 See the License for the specific language governing permissions and
 limitations under the License.
 */
-package markets.tradables.orders
-
-import markets.tradables.Tradable
+package markets.tradables
 
 
 trait Quantity {
